+++ lib/ext/Appointment.dart
@@ -14,14 +14,20 @@ import 'package:sprintf/sprintf.dart';
 
 extension AppointmentExt on Appointment {
 
-  String? get displayScheduleTime =>
-    AppointmentTimeSlotExt.getDisplayScheduleTime(startTimeUtc, endTimeUtc);
+  String? get displayLongScheduleTime =>
+    AppointmentTimeSlotExt.getLongDisplayScheduleTime(startTimeUtc, endTimeUtc);
+
+  String? get displayShortScheduleTime =>
+    AppointmentTimeSlotExt.getShortDisplayScheduleTime(startTimeUtc, endTimeUtc);
 
   int? get startMinutesSinceMidnightUtc =>
     AppointmentTimeSlotExt.getStartMinutesSinceMidnightUtc(startTimeUtc);
 
   String get displayProviderName =>
     this.provider?.name ?? Localization().getStringEx('model.academics.appointment.default.provider.label', 'MyMcKinley');
+
+  String? get displayType =>
+    appointmentTypeToDisplayString(type, provider: provider);
 
   String? get category =>
     sprintf(Localization().getStringEx('model.academics.appointment.category.label.format', '%s Appointments'), [displayProviderName]).toUpperCase();
@@ -117,19 +123,36 @@ extension AppointmentTimeSlotExt on AppointmentTimeSlot {
   DateTime? get startTime => startTimeUtc?.toUniOrLocal();
   DateTime? get endTime => endTimeUtc?.toUniOrLocal();
 
-  String? get displayScheduleTime =>
-    getDisplayScheduleTime(startTimeUtc, endTimeUtc);
+  String? get displayLongScheduleTime =>
+    getLongDisplayScheduleTime(startTimeUtc, endTimeUtc);
 
-  static String? getDisplayScheduleTime(DateTime? startTimeUtc, DateTime? endTimeUtc) {
+  String? get displayShortScheduleTime =>
+    getShortDisplayScheduleTime(startTimeUtc, endTimeUtc);
+
+  static String? getLongDisplayScheduleTime(DateTime? startTimeUtc, DateTime? endTimeUtc) {
     if (startTimeUtc != null) {
       if (endTimeUtc != null) {
-        //AppDateTime().getDeviceTimeFromUtcTime(startTime)
-        String startTimeStr = DateFormat('EEEE, MMMM d, yyyy hh:mm').format(startTimeUtc.toUniOrLocal());
-        String endTimeStr = DateFormat('hh:mm aaa').format(endTimeUtc.toUniOrLocal());
+        String startTimeStr = DateFormat('EEEE, MMMM d, yyyy h:mm').format(startTimeUtc.toUniOrLocal());
+        String endTimeStr = DateFormat('h:mm aaa').format(endTimeUtc.toUniOrLocal());
         return "$startTimeStr - $endTimeStr";
       }
       else {
-        return DateFormat('EEEE, MMMM d, yyyy hh:mm aaa').format(startTimeUtc.toUniOrLocal());
+        return DateFormat('EEEE, MMMM d, yyyy h:mm aaa').format(startTimeUtc.toUniOrLocal());
+      }
+    }
+    return null;
+  }
+
+  static String? getShortDisplayScheduleTime(DateTime? startTimeUtc, DateTime? endTimeUtc) {
+    if (startTimeUtc != null) {
+      if (endTimeUtc != null) {
+        //AppDateTime().getDeviceTimeFromUtcTime(startTime)
+        String startTimeStr = DateFormat('EEE, MMM d, yyyy h:mm').format(startTimeUtc.toUniOrLocal());
+        String endTimeStr = DateFormat('h:mm aaa').format(endTimeUtc.toUniOrLocal());
+        return "$startTimeStr-$endTimeStr";
+      }
+      else {
+        return DateFormat('EEE, MMM d, yyyy h:mm aaa').format(startTimeUtc.toUniOrLocal());
       }
     }
     return null;
@@ -145,20 +168,15 @@ extension AppointmentTimeSlotExt on AppointmentTimeSlot {
 ///////////////////////////////
 /// AppointmentType
 
-String? appointmentTypeToDisplayString(AppointmentType? type) {
+String? appointmentTypeToDisplayString(AppointmentType? type, { AppointmentProvider? provider }) {
   switch (type) {
     case AppointmentType.in_person:
       return Localization().getStringEx('model.academics.appointment.type.in_person.label', 'In Person');
     case AppointmentType.online:
-      return Localization().getStringEx('model.academics.appointment.type.online.label', 'Telehealth');
+      return (provider?.name == AppointmentProviderExt.mcKinleyName) ?
+        Localization().getStringEx('model.academics.appointment.type.telehealth.label', 'Telehealth') :
+        Localization().getStringEx('model.academics.appointment.type.online.label', 'Online');
     default:
       return null;
-  }
-}
-
-String appointment2TypeDisplayString(AppointmentType _appointmentType) {
-  switch (_appointmentType) {
-    case AppointmentType.in_person: return Localization().getStringEx('model.academics.appointment2.type.in_person.label', 'In Person');
-    case AppointmentType.online: return Localization().getStringEx('model.academics.appointment2.type.online.label', 'Online');
   }
 }