+++ lib/ui/appointments/AppointmentScheduleTimePanel.dart
@@ -229,7 +229,7 @@ class _AppointmentScheduleTimePanelState extends State<AppointmentScheduleTimePa
       return Container();
     }
     else {
-      String? currentDateString = AppointmentTimeSlotExt.getDisplayScheduleTime(widget.sourceAppointment?.startTimeUtc, widget.sourceAppointment?.endTimeUtc) ??
+      String? currentDateString = AppointmentTimeSlotExt.getLongDisplayScheduleTime(widget.sourceAppointment?.startTimeUtc, widget.sourceAppointment?.endTimeUtc) ??
         Localization().getStringEx('panel.appointment.reschedule.time.label.unknown', 'Unknown');
       
       return Padding(padding:EdgeInsets.only(left: 16, right: 16, top: 16), child: